revised to correclty build Omega on Perlmutter and Frontier.
* creates scripts that prerun module commands in the build direcotry:
  omega_build.sh, omega_run.sh and omega_ctest.sh
* YAKL library can be automatically built using HIP or CUDA
* updates documents for cmake-based build.

+++ components/omega/OmegaBuild.cmake
@@ -1,5 +1,5 @@
 ###########################
-# Internal variables      #
+# Build Control Variables #
 ###########################
 
 set(OMEGA_PROJECT_NAME            "OmegaOceanModel")
@@ -17,8 +17,10 @@ set(E3SM_CIME_ROOT ${OMEGA_SOURCE_DIR}/../../cime)
 set(E3SM_CIMECONFIG_ROOT ${OMEGA_SOURCE_DIR}/../../cime_config)
 
 ###########################
-# Public variables        #
+# Macros                  #
 ###########################
+
+# set build control variables used for both e3sm build and standalone build
 macro(setup_common_variables)
 
   option(OMEGA_DEBUG "Turn on error message throwing (default OFF)." OFF)
@@ -27,23 +29,13 @@ macro(setup_common_variables)
     set(OMEGA_CXX_FLAGS "")
   endif()
 
-  if(NOT DEFINED OMEGA_INCLUDE_DIRECTORIES)
-    set(OMEGA_INCLUDE_DIRECTORIES "")
-  endif()
-
-  if(NOT DEFINED OMEGA_LINK_DIRECTORIES)
-    set(OMEGA_LINK_DIRECTORIES "")
-  endif()
-
   if(NOT DEFINED OMEGA_LINK_OPTIONS)
     set(OMEGA_LINK_OPTIONS "")
   endif()
 
 endmacro()
 
-###########################
-# Preset Standalone build #
-###########################
+# Collect machine and compiler info from CIME
 macro(preset)
 
   find_package (Python COMPONENTS Interpreter)
@@ -53,36 +45,49 @@ macro(preset)
   endif()
 
   set(_TMP_CMAKE_FILE ${CMAKE_CURRENT_BINARY_DIR}/_Omega.cmake)
+  set(_PY_OPTS "-p;${E3SM_CIME_ROOT};-o;${_TMP_CMAKE_FILE}")
 
   if(DEFINED OMEGA_CIME_COMPILER)
-    set(_TMP_COMPILER "-c ${OMEGA_CIME_COMPILER}")
-  else()
-    set(_TMP_COMPILER "")
+    list(APPEND _PY_OPTS "-c" "${OMEGA_CIME_COMPILER}")
   endif()
 
-  execute_process(COMMAND ${Python_EXECUTABLE} gen_machine_info.py
-    -p ${E3SM_CIME_ROOT} -o ${_TMP_CMAKE_FILE} ${_TMP_COMPILER}
+  if(DEFINED OMEGA_CIME_MACHINE)
+    list(APPEND _PY_OPTS "-m" "${OMEGA_CIME_MACHINE}")
+  endif()
+
+  if(OMEGA_BUILD_TYPE STREQUAL "Debug")
+    list(APPEND _PY_OPTS "-d")
+  endif()
+
+  execute_process(COMMAND ${Python_EXECUTABLE} create_scripts.py ${_PY_OPTS}
     OUTPUT_QUIET ERROR_QUIET
     WORKING_DIRECTORY ${OMEGA_SOURCE_DIR}
     OUTPUT_VARIABLE _MACHINE_INFO)
 
   include(${_TMP_CMAKE_FILE})
-  file(REMOVE ${_TMP_CMAKE_FILE})
+  if(OMEGA_BUILD_TYPE STREQUAL "Release")
+    file(REMOVE ${_TMP_CMAKE_FILE})
+  endif()
 
-  if(NOT OMEGA_CXX_COMPILER)
+  if(OMEGA_CXX_COMPILER)
+    find_program(_OMEGA_CXX_COMPILER ${OMEGA_CXX_COMPILER})
+  else()
     if (MPILIB STREQUAL "mpi-serial")
-      set(OMEGA_CXX_COMPILER ${SCXX})
+      find_program(_OMEGA_CXX_COMPILER ${SCXX})
     else()
-      set(OMEGA_CXX_COMPILER ${MPICXX})
+      find_program(_OMEGA_CXX_COMPILER ${MPICXX})
     endif()
   endif()
 
+  set(OMEGA_CXX_COMPILER ${_OMEGA_CXX_COMPILER})
   set(CMAKE_CXX_COMPILER ${OMEGA_CXX_COMPILER})
+
   message(STATUS "OMEGA_CXX_COMPILER = ${OMEGA_CXX_COMPILER}")
   message(STATUS "MPI_EXEC = ${MPI_EXEC}")
 
 endmacro()
 
+# set build-control-variables for standalone build
 macro(setup_standalone_build)
 
   setup_common_variables()
@@ -110,6 +115,7 @@ macro(setup_standalone_build)
 
 endmacro()
 
+# set build-control-variables for e3sm build
 macro(setup_e3sm_build)
 
   setup_common_variables()
@@ -117,10 +123,13 @@ macro(setup_e3sm_build)
   set(OMEGA_BUILD_TYPE ${E3SM_DEFAULT_BUILD_TYPE})
   set(E3SM_EXTERNALS_ROOT ${E3SM_SOURCE_DIR}/../externals)
 
+  set(OMEGA_CXX_COMPILER ${CMAKE_CXX_COMPILER})
 
   #TODO: set OMEGA_ARCH according to E3SM variables
   set(OMEGA_ARCH "NOT_DEFINED")
   set(OMEGA_BUILD_MODE "E3SM")
+
+  message(STATUS "OMEGA_CXX_COMPILER = ${OMEGA_CXX_COMPILER}")
 
 endmacro()
 
@@ -133,18 +142,23 @@ macro(update_variables)
   # Set the build type
   set(CMAKE_BUILD_TYPE ${OMEGA_BUILD_TYPE})
 
-  if (NETCDF_PATH)
-    list(APPEND OMEGA_INCLUDE_DIRECTORIES "${NETCDF_PATH}/include")
-    list(APPEND OMEGA_LINK_DIRECTORIES "${NETCDF_PATH}/lib")
-    list(APPEND OMEGA_LINK_OPTIONS "-lnetcdf")
+  # Set compiler and linker flags
+  if (CXXFLAGS)
+    separate_arguments(_CXXFLAGS NATIVE_COMMAND ${CXXFLAGS})
+    list(APPEND OMEGA_CXX_FLAGS ${_CXXFLAGS})
   endif()
 
-  if (PNETCDF_PATH)
-    list(APPEND OMEGA_INCLUDE_DIRECTORIES "${PNETCDF_PATH}/include")
-    list(APPEND OMEGA_LINK_DIRECTORIES "${PNETCDF_PATH}/lib")
-    list(APPEND OMEGA_LINK_OPTIONS "-lpnetcdf")
+  if (LDFLAGS)
+    separate_arguments(_LDFLAGS NATIVE_COMMAND ${LDFLAGS})
+    list(APPEND OMEGA_LINK_OPTIONS ${_LDFLAGS})
   endif()
 
+  if (SLIBS)
+    separate_arguments(_SLIBS NATIVE_COMMAND ${SLIBS})
+    list(APPEND OMEGA_LINK_OPTIONS ${_SLIBS})
+  endif()
+
+  # check if MPI is supported
   string(CONCAT _TestMPISource
     "#include \"mpi.h\"\n"
     "int main(int argc, char* argv[])\n"
@@ -156,7 +170,9 @@ macro(update_variables)
   execute_process(
 	COMMAND ${OMEGA_CXX_COMPILER} -c ${_TestMPISrcFile} -o ${_TestMPIObjFile}
     OUTPUT_QUIET ERROR_QUIET
-    RESULT_VARIABLE _MPI_TEST_RESULT)
+    RESULT_VARIABLE _MPI_TEST_RESULT
+    OUTPUT_VARIABLE _MPI_TEST_OUTPUT
+    ERROR_VARIABLE _MPI_TEST_ERROR)
 
   if(OMEGA_BUILD_TYPE EQUAL Release)
 	file(REMOVE ${_TestMPISrcFile})
@@ -167,9 +183,10 @@ macro(update_variables)
     if (_MPI_TEST_RESULT MATCHES "^[-]?[0-9]+$")
       find_package(MPI)
       if(MPI_CXX_FOUND)
-        list(APPEND OMEGA_INCLUDE_DIRECTORIES "${MPI_CXX_INCLUDE_DIRS}")
-        list(APPEND OMEGA_LINK_DIRECTORIES "${MPI_CXX_INCLUDE_DIRS}/../lib")
-        list(APPEND OMEGA_LINK_OPTIONS "-lmpi")
+        list(APPEND OMEGA_CXX_FLAGS "-I${MPI_CXX_INCLUDE_DIRS}")
+        list(APPEND OMEGA_LINK_OPTIONS
+          "-L${MPI_CXX_INCLUDE_DIRS}/../lib" "-lmpi"
+        )
       else()
         message(FATAL_ERROR "MPI is not found" )
       endif()
@@ -178,50 +195,67 @@ macro(update_variables)
     endif()
   endif()
 
-  if(USE_CUDA)
-    list(APPEND OMEGA_CXX_FLAGS ${CUDA_FLAGS})
-
-  elseif(USE_HIP)
-    list(APPEND OMEGA_CXX_FLAGS ${HIP_FLAGS})
-
-  elseif(USE_SYCL)
-    list(APPEND OMEGA_CXX_FLAGS ${SYCL_FLAGS})
-
-  else()
-    list(APPEND OMEGA_CXX_FLAGS ${CXXFLAGS})
-
-    if(CXX_LIBS)
-      list(APPEND OMEGA_LINK_OPTIONS  ${CXX_LIBS})
-    endif()
-
-  endif()
-
-  message(STATUS "OMEGA_CXX_FLAGS           = ${CMAKE_CXX_FLAGS}")
-  message(STATUS "OMEGA_INCLUDE_DIRECTORIES = ${OMEGA_INCLUDE_DIRECTORIES}")
-  message(STATUS "OMEGA_LINK_DIRECTORIES    = ${OMEGA_LINK_DIRECTORIES}")
+  message(STATUS "OMEGA_CXX_FLAGS           = ${OMEGA_CXX_FLAGS}")
   message(STATUS "OMEGA_LINK_OPTIONS        = ${OMEGA_LINK_OPTIONS}")
 
   if(OMEGA_INSTALL_PREFIX)
     set(CMAKE_INSTALL_PREFIX ${OMEGA_INSTALL_PREFIX})
   endif()
 
-  if(DEFINED OMEGA_ARCH)
+  # Check if CUDA or HIP is supported
+  if((NOT DEFINED OMEGA_ARCH) OR (OMEGA_ARCH STREQUAL "NOT_DEFINED"))
 
-    if(OMEGA_ARCH STREQUAL "NOT_DEFINED")
-      set(YAKL_ARCH "")
+    execute_process(
+	  COMMAND ${OMEGA_CXX_COMPILER} --version
+      RESULT_VARIABLE _CXX_VER_RESULT
+      OUTPUT_VARIABLE _CXX_VER_OUTPUT)
 
+    if (_CXX_VER_RESULT EQUAL 0)
+
+      string(REGEX MATCH "HIP|hip"       _HIP_CHECK "${_CXX_VER_OUTPUT}")
+      string(REGEX MATCH "AMD|amd"       _AMD_CHECK "${_CXX_VER_OUTPUT}")
+      string(REGEX MATCH "NVCC|nvcc"     _NVCC_CHECK "${_CXX_VER_OUTPUT}")
+      string(REGEX MATCH "NVIDIA|nvidia" _NVIDIA_CHECK "${_CXX_VER_OUTPUT}")
+
+      if(_HIP_CHECK AND _AMD_CHECK)
+        set(OMEGA_ARCH "HIP")
+
+      elseif(_NVCC_CHECK AND _NVIDIA_CHECK)
+        set(OMEGA_ARCH "CUDA")
+
+      else()
+        set(OMEGA_ARCH "")
+
+      endif()
     else()
-      set(YAKL_ARCH ${OMEGA_ARCH})
+        set(OMEGA_ARCH "")
+
+    endif()
+  endif()
+
+  if(OMEGA_ARCH STREQUAL "CUDA")
+    enable_language(CUDA)
+
+  elseif(OMEGA_ARCH STREQUAL "HIP")
+    enable_language(HIP)
+
+  elseif(USE_CUDA)
+    set(OMEGA_ARCH "CUDA")
+    enable_language(CUDA)
+
+  elseif(USE_HIP)
+    set(OMEGA_ARCH "HIP")
+    enable_language(HIP)
+
+  endif()
+
+  set(YAKL_ARCH "${OMEGA_ARCH}")
+
+  if(YAKL_ARCH)
 
       if(OMEGA_${YAKL_ARCH}_FLAGS)
         set(YAKL_${YAKL_ARCH}_FLAGS ${OMEGA_${YAKL_ARCH}_FLAGS})
       endif()
-
-    endif()
-
-  else()
-
-    set(YAKL_ARCH "")
 
   endif()
 
@@ -269,10 +303,14 @@ macro(wrap_outputs)
 
   if(OMEGA_INSTALL_PREFIX)
 
-    install(TARGETS ${OMEGA_LIB_NAME} LIBRARY DESTINATION "${OMEGA_INSTALL_PREFIX}/lib")
+    install(TARGETS ${OMEGA_LIB_NAME}
+      LIBRARY DESTINATION "${OMEGA_INSTALL_PREFIX}/lib"
+    )
 
     if(OMEGA_BUILD_EXECUTABLE)
-      install(TARGETS ${OMEGA_EXE_NAME} RUNTIME DESTINATION "${OMEGA_INSTALL_PREFIX}/bin")
+      install(TARGETS ${OMEGA_EXE_NAME}
+        RUNTIME DESTINATION "${OMEGA_INSTALL_PREFIX}/bin"
+      )
     endif()
 
   endif()
